Upload and parse data for JAMES

Client side upload of a JSON file, string or URL with BDS data, checks the data, stores its contents as a tibble with a person attribute on host, and returns an object of class httr::response() that contains the results of the request.

upload_txt(
  txt,
  host = "http://localhost",
  mod = character(0),
  format = "2.0",
  schema = NULL,
  verbose = FALSE
)

Arguments

txt

Data set. Could be a URL, local file, JSON string or JSON object. See bdsreader::read_bds().

host

String with the host. The default is "http://localhost".

mod

String with the module path, if any. The default is "".

format

String. JSON data schema version number. There are currently three schemas supported: "1.0", "1.1" and "2.0". Formats "1.0" and "1.1" are included for backward compatibility only. Use format = "2.0" for new applications.

schema

A file name (optionally including the path) with the JSON validation schema. The schema argument overrides format. The function extracts the version number for the basename, and overwrites the format argument by version number.

verbose

Logical. Print diagnostic information of POST request to console.

Value

An object of class httr::response()

Details

JSON format: See https://stefvanbuuren.name/jamesdocs/getting-data-into-james.html for the specification of the JSON format.

User agent: The function upload_txt() searches for an object called ua on the search list. The ua object is an optional user agent, a request that identifies yourself to the API. For example, run httr::user_agent("https://github.com/myaccount") (with myaccount replaced by your github user name) before calling upload_txt(). See https://httr.r-lib.org/articles/api-packages.html for details. Setting the user agent is not required.

Append "/json" to path and set query = "auto_unbox=TRUE&force=TRUE" to obtain a partial JSON representation of the S4 class individual. At present, it is not possible to rebuild the S4 class individual from its JSON representation because the S4 class depends on environments, and these are not converted to JSON. Warning: The S4 class individual is an internal format that is in development. It is likely to change, so don't build applications based on this data structure. If you need components from the internal structure (e.g. Z-scores, brokenstick estimates) it is better to develop a dedicated API for obtaining these.

Examples

library(httr)

host <- "http://localhost"
url <- paste(host, "ocpu/library/jamesdemodata",
  "extdata/bds_v2.0/smocc/Laura_S.json", sep = "/")
fn <- system.file("extdata", "bds_v2.0", "smocc", "Laura_S.json",
 package = "jamesdemodata", mustWork = TRUE)
js1 <- readLines(url)
#> Warning: URL 'http://localhost/ocpu/library/jamesdemodata/extdata/bds_v2.0/smocc/Laura_S.json': status was 'Couldn't connect to server'
#> Error in file(con, "r"): cannot open the connection to 'http://localhost/ocpu/library/jamesdemodata/extdata/bds_v2.0/smocc/Laura_S.json'
js2 <- jsonlite::toJSON(jsonlite::fromJSON(url), auto_unbox = TRUE)
#> Error in open.connection(con, "rb"): Failed to connect to localhost port 80: Connection refused

# upload JSON file
r1 <- upload_txt(fn, host)
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to localhost port 80: Connection refused
identical(status_code(r1), 201L)
#> Error in status_code(r1): object 'r1' not found

# upload JSON string
r2a <- upload_txt(js1, host)
#> Error in file.exists(txt[1L]): object 'js1' not found
identical(status_code(r2a), 201L)
#> Error in status_code(r2a): object 'r2a' not found
r2b <- upload_txt(js2, host)
#> Error in file.exists(txt[1L]): object 'js2' not found
identical(status_code(r2b), 201L)
#> Error in status_code(r2b): object 'r2b' not found

# upload JSON from external URL.
r3 <- upload_txt(url, host, format = "1.0")
#> Error in open.connection(con, open = mode): Failed to connect to localhost port 80: Connection refused
identical(status_code(r3), 201L)
#> Error in status_code(r3): object 'r3' not found
# browseURL(get_url(r3))

r <- GET(paste(get_url(r3), "json", sep = "/"))
#> Error in is.response(x): object 'r3' not found
r
#> Error in eval(expr, envir, enclos): object 'r' not found